requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Gradio UI
gradio==4.16.0
huggingface-hub==0.20.3

# LangChain and LangGraph
langchain>=0.1.0,<0.2.0
langchain-openai==0.1.1
langchain-core>=0.1.0
langchain-community>=0.0.18
langgraph>=0.0.20

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Environment and configuration
python-dotenv==1.0.0

# Twilio (WhatsApp)
twilio==8.13.0

# HTTP requests (for HubSpot)
requests==2.31.0

# Vector database and RAG (disabled - requires Visual C++)
# chromadb==0.4.22

# Document processing (disabled - requires chromadb)
# pymupdf==1.23.21
# python-docx==1.1.0
# unstructured==0.11.8

# Scheduler
apscheduler==3.10.4

# OpenAI
openai==1.10.0

# Data validation
pydantic==2.5.3

# Typing extensions
typing-extensions==4.9.0
